+++ lib/src/utils/extensions/double_extensions.dart
