+++ lib/src/utils/extensions/string_extensions.dart
@@ -1,4 +1,4 @@
-import '../../utils.dart';
+import '../regex/get_utils.dart';
 
 extension GetStringUtils on String {
   bool get isNum => GetUtils.isNum(this);
