+++ lib/utils.dart
@@ -1,5 +1,4 @@
-export 'src/utils/context_extensions/extensions.dart';
+export 'src/utils/extensions/export.dart';
 export 'src/utils/queue/get_queue.dart';
 export 'src/utils/platform/platform.dart';
 export 'src/utils/regex/get_utils.dart';
-export 'src/utils/regex/get_utils_extensions.dart';
